msm8937-common: check_features: fix NFC removal
* Move /vendor mounting to script, since in releasetools it gives a build error

+++ configs/check_features.sh
@@ -4,8 +4,11 @@ sku=`getprop ro.boot.hardware.sku`
 device=`getprop ro.boot.device`
 radio=`getprop ro.boot.radio`
 
-if [ "$device" == "cedric" ] || [ "$device" == "owens" ] || [ "$device" == "perry" ]; then
-    # owens, perry and cedric don't have e-compasses
+umount /vendor
+mount /vendor
+
+if [ "$radio" == "APAC" ] || [ "$radio" == "EMEA" ] && [ "$device" != "cedric" ] && [ "$device" != "owens" ] && [ "$device" != "perry" ]; then
+    # On montana APAC and EMEA radios don't have e-compasses
     rm /vendor/etc/permissions/android.hardware.sensor.compass.xml
 fi
 
@@ -13,15 +16,10 @@ if [ "$device" == "montana" ]; then
     if [ "$radio" != "APAC" ] && [ "$radio" != "EMEA" ]; then
         # On montana radios other than APAC and EMEA don't have NFC chips
         rm /system/etc/permissions/com.android.nfc_extras.xml
-        rm /vendor/etc/libnfc-nci.conf
-        rm /vendor/etc/libnfc-nxp.conf
         rm /vendor/etc/permissions/android.hardware.nfc.xml
         rm /vendor/etc/permissions/android.hardware.nfc.hce.xml
-        rm /vendor/firmware/libpn548ad_fw.so
         rm -r /system/app/NfcNci
     fi
-    if [ "$radio" == "APAC" ] || [ "$radio" == "EMEA" ]; then
-        # On montana APAC and EMEA radios don't have e-compasses
-        rm /vendor/etc/permissions/android.hardware.sensor.compass.xml
-    fi
 fi
+
+umount /vendor
